issue#168 - If connection is refused - write it to login window
Problem: if connection is broken or refused, the message about wrong
credentials will be shown. This is inappropriate behavior

Analysis:
The problem is in LoginServiceImpl::login method. If HttpResponseHolder
instance returned empty, the message about wrong creds will be shown.

Solution: add the check for response code. If we get the request
timeout, the 408 code will be returned. On the other hand, if we get
some error during connection, response code in HttpResponseHolder will
be equals to 0. In these cases we should show connection error message.

Also provided some refactoring