bug#184 - shadow project must not provide any operation except cloning
Problem: if user selects shadow project and pushes it, the push
application state will be run but never ends. Same problem could be
observed for another Git operations.

Analysis: If we start to push shadow project, the empty list will be
passed to GitService::push method with new instance of
PushProgressListener. Push operation will be rejected because of empty
list, but push application state is never changes back, because in this
case ProgressListener::onFinish method is never invokes.

The global problem is that we could invoke every Git operation even if
list of proper projects, that is returned by
MainWindowController::getProjectsClonedAndWithoutConflicts, is empty.

Solution: add List::isEmpty check for Git operations to prevent changing application status if operation could not be fired.