issue #200 - status bar modification
Problem 1: status bar progress indicator and statuses should be shown on
the right part of the screen

Solution 1: provide CENTER_RIGHT alignment for the StatusBar in the
initUI() method

Problem 2: StatusBar doesn't show progress during group loading

Analysis 2: ApplicationState.REFRESH_PROJECTS turns on only during
refreshing projects via "Refresh Projects" button on the toolbar. It is
better to change application state in
ProjectServiceImpl.loadProjects(Group) method. In this case this state
will be always shown, if someone reloads objects

Solution 2: move invocation of the StateService::stateON and
StateService::stateOFF methods from ProjectList::refreshLoadProjects to
ProjectServiceImpl::loadProjects

Problem 3: Console could be resized if we drag the top border of the
StatusBar

Analysis 3: StatusBar has been added to the &lt;center&gt; part of the
ModularContainer, so that it takes the border of the SplitPanel

Solution: move StatusBar to the &lt;bottom&gt; part of the ModularContainer