bug in GitServiceImpl::switchTo method
Problem: switch branch state is never ends if we start it after branch
creation

Analysis: the root cause is the nullable ProgressListener.
If we pass the null ProgressListener value to the
GitServiceImpl::switchTo method, it creates the instance of
EmptyProgressListener. This stub listener doesn't switched off any
status in the onFinish() method unlike in
OperationStateListener::onFinish method.

This is a regression caused by #204

Solution: switch off the SWITCH_STATE in the finally block in GitServiceImpl::runSwitchAction method if this state still active