#242 - gitlab URL forward slash issue
Problem: If it gitlab URL has forward slash in the end of URL, connection doesn't work.
E.g.
https://git.openearth.community/ doesn't work.
https://git.openearth.community Looks like this work

Analysis: the problem is in URL validator (see URLManager.isURLValid() method). The regular expression does not contain the back slash in the end.

Solution: add the back slash handling in the end of the string to the regular expression